+++ bibliotecaComunitaria/src/biblioteca/arquivo/Editor.java
@@ -1,14 +1,8 @@
 package biblioteca.arquivo;
 
 import biblioteca.livros.*;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
+import java.io.*;
+
 
 /* @author Luam */
 public class Editor {
