+++ bibliotecaComunitaria/src/biblioteca/arquivo/Escritor.java
@@ -3,11 +3,10 @@ package biblioteca.arquivo;
 import biblioteca.livros.Estante;
 import biblioteca.livros.Livro;
 import biblioteca.pessoas.Autor;
-import java.io.BufferedWriter;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
+import biblioteca.pessoas.Cliente;
+import biblioteca.pessoas.Funcionario;
+import java.io.*;
+import java.util.ArrayList;
 
 /* @author Luam */
 
@@ -58,7 +57,7 @@ public class Escritor {
             }   
         }
     }
-    public static void escreveLivro(Livro livro) {
+    public static void escreverLivro(Livro livro) {
         BufferedWriter bw = null;
         String linha = livro.getAutor().getNome() + "," + livro.getTitulo() + "," + livro.getNumPaginas() + "," +
                 livro.getISBN() + "," + livro.getGenero()  + "," + livro.getEditora() + "," + 
@@ -83,4 +82,54 @@ public class Escritor {
             }   
         }
     }
+    
+    public static void escreverCliente(Cliente cliente) {
+        BufferedWriter bw = null;
+        String linha = cliente.getNome() + "," + cliente.getNascimento() + "," + cliente.getTelefone() + "," + cliente.getEnd().getRua() + "," + cliente.getEnd().getBairro() + "," + cliente.getEnd().getCep() + "," + cliente.getEnd().getCidade() + "," + cliente.getEnd().getEstado();
+        try {
+            bw = new BufferedWriter(new FileWriter("clientes.csv", true));
+            PrintWriter pw = new PrintWriter(bw);
+            pw.println(linha);
+            pw.flush();
+            pw.close();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (bw != null) {
+                try {
+                    bw.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }   
+        }
+    }
+    
+    public static void escreverFuncionario(Funcionario funcionario) {
+        BufferedWriter bw = null;
+        String linha = funcionario.getNome() + "," + funcionario.getNascimento() + "," + funcionario.getTelefone() + ","+ funcionario.getSalario() + "," + funcionario.getCargo() + "," + funcionario.getEnd().getRua() + "," + funcionario.getEnd().getBairro() + "," + funcionario.getEnd().getCep() + "," + funcionario.getEnd().getCidade() + "," + funcionario.getEnd().getEstado();
+        try {
+            bw = new BufferedWriter(new FileWriter("funcionarios.csv", true));
+            PrintWriter pw = new PrintWriter(bw);
+            pw.println(linha);
+            pw.flush();
+            pw.close();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            if (bw != null) {
+                try {
+                    bw.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }   
+        }
+    }
+    
+    
 }
