+++ bibliotecaComunitaria/src/biblioteca/arquivo/Removedor.java
@@ -2,14 +2,7 @@ package biblioteca.arquivo;
 
 import biblioteca.livros.Acervo;
 import biblioteca.livros.Livro;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
+import java.io.*;
 
 /* @author Luam */
