+++ bibliotecaComunitaria/src/biblioteca/arquivo/TesteClientes.java
@@ -3,7 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package biblioteca.pessoas;
+package biblioteca.arquivo;
+import biblioteca.pessoas.Cliente;
+import biblioteca.pessoas.Funcionario;
 import java.util.ArrayList;
 
 /**
@@ -15,28 +17,33 @@ public class TesteClientes {
     public static void main(String[] args){
         ArrayList<Cliente> clientes = new ArrayList();
         
-        clientes = LeituraPessoas.leitorClientes();
-        LeituraPessoas.imprimir(clientes);
+        clientes = Leitor.leitorClientes();
+        imprimir(clientes);
 
         System.out.println("\n--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------\n");
         
         ArrayList<Funcionario> funcionarios = new ArrayList();
 
-        funcionarios = LeituraPessoas.leitorFuncionarios();
-        LeituraPessoas.imprimir(funcionarios);
+        funcionarios = Leitor.leitorFuncionarios();
+        imprimir(funcionarios);
         
         /*Cliente cliente = new Cliente();
         cliente.cadastro("Maria", "12/12/2000", "99999-9999", "Rua 1", "Campo Grande", "22222-222", "Rio de Janeiro", "RJ");
-        LeituraPessoas.escreverCliente(cliente);
-        clientes = LeituraPessoas.leitorClientes();
+        Escritor.escreverCliente(cliente);
+        clientes = Leitor.leitorClientes();
         LeituraPessoas.imprimir(clientes);*/
         
         /*Funcionario funcionario = new Funcionario();
         funcionario.cadastro("Maria", "12/12/2000", "99999-9999", 4000, "gerente", "Rua 1", "Campo Grande", "22222-222", "Rio de Janeiro", "RJ");
         
-        LeituraPessoas.escreverFuncionario(funcionario);
-        funcionarios = LeituraPessoas.leitorFuncionarios();
-        LeituraPessoas.imprimir(funcionarios);*/
+        Escritor.escreverFuncionario(funcionario);
+        funcionarios = Leitor.leitorFuncionarios();
+        imprimir(funcionarios);*/
     }
-
+    
+    public static void imprimir(ArrayList a){
+        for (Object i: a){
+           System.out.println(i);
+        }
+    }
 }
